backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0
xgboost==2.0.3
matplotlib==3.8.2
seaborn==0.13.1
scipy==1.12.0
